Time: 40 ms (52.67%) | Memory: 105.7 MB (92.57%) - LeetSync

+++ 1966-frequency-of-the-most-frequent-element/frequency-of-the-most-frequent-element.cpp
@@ -14,6 +14,6 @@ public:
         ans=max(ans,static_cast<long>(r-l+1));
         r++;
         }
-        return static_cast<long>(ans);
+        return ans;
     }
 };
